extract breadcrumbs UI element
